refactor: modularize CLI commands and unify scripts
