+++ doc_ai/cli/__init__.py
@@ -1,26 +1,22 @@
 """CLI orchestrator for AI document analysis pipeline."""
 from __future__ import annotations
 
-from pathlib import Path
-from typing import Optional
+import logging
 import os
 import sys
 import traceback
-import warnings
-import logging
+import importlib
+from pathlib import Path
+from typing import Optional
 
 import typer
-from typer.completion import completion_init, get_completion_script, Shells
+from dotenv import find_dotenv, load_dotenv
 from rich.console import Console
-from rich.table import Table
-from dotenv import load_dotenv, set_key, find_dotenv
-from .interactive import interactive_shell, get_completions
+from typer.completion import Shells, completion_init, get_completion_script
 
-# Ensure project root is first on sys.path when running as a script.
-if __package__ in (None, ""):
-    sys.path[0] = str(Path(__file__).resolve().parent.parent)
-
+from doc_ai import __version__
 from doc_ai.converter import OutputFormat, convert_path
+from .interactive import get_completions, interactive_shell
 from .utils import (
     EXTENSION_MAP,
     analyze_doc,
@@ -31,7 +27,9 @@ from .utils import (
     validate_doc,
 )
 
-from doc_ai import __version__
+# Ensure project root is first on sys.path when running as a script.
+if __package__ in (None, ""):
+    sys.path[0] = str(Path(__file__).resolve().parent.parent)
 
 ENV_FILE = find_dotenv(usecwd=True, raise_error_if_not_found=False) or ".env"
 
@@ -167,267 +165,23 @@ def completion(shell: Shells):
     typer.echo(script)
 
 
-@app.command()
-def config(
-    verbose: bool = typer.Option(
-        None, "--verbose/--no-verbose", help="Toggle verbose error output"
-    ),
-    set_vars: list[str] = typer.Option(
-        None,
-        "--set",
-        help="Set VAR=VALUE pairs to update environment configuration",
-        metavar="VAR=VALUE",
-    ),
-) -> None:
-    """Show or update runtime configuration."""
-    if verbose is not None:
-        SETTINGS["verbose"] = verbose
-    if set_vars:
-        env_path = Path(ENV_FILE)
-        env_path.touch(exist_ok=True)
-        env_path.chmod(0o600)
-        for item in set_vars:
-            try:
-                key, value = item.split("=", 1)
-            except ValueError as exc:  # pragma: no cover - handled by typer
-                raise typer.BadParameter("Use VAR=VALUE syntax") from exc
-            os.environ[key] = value
-            set_key(str(env_path), key, value, quote_mode="never")
-            env_path.chmod(0o600)
-    console.print("Current settings:")
-    console.print(f"  verbose: {SETTINGS['verbose']}")
-    defaults = load_env_defaults()
-    if defaults:
-        table = Table("Variable", "Current", "Default")
-        for var, default in sorted(defaults.items()):
-            table.add_row(var, os.getenv(var, "") or "-", default or "-")
-        console.print(table)
+# Register subcommands implemented in dedicated modules.
+from . import analyze as analyze_cmd
+from . import config as config_cmd
+from . import convert as convert_cmd
+from . import embed as embed_cmd
+pipeline_cmd = importlib.import_module("doc_ai.cli.pipeline")
+from . import validate as validate_cmd
 
+app.add_typer(config_cmd.app, name="config")
+app.add_typer(convert_cmd.app, name="convert")
+app.add_typer(validate_cmd.app, name="validate")
+app.add_typer(analyze_cmd.app, name="analyze")
+app.add_typer(embed_cmd.app, name="embed")
+app.add_typer(pipeline_cmd.app, name="pipeline")
 
-@app.command()
-def convert(
-    source: str = typer.Argument(
-        ..., help="Path or URL to raw document or folder"
-    ),
-    format: list[OutputFormat] = typer.Option(
-        None,
-        "--format",
-        "-f",
-        help="Desired output format(s). Can be passed multiple times.",
-    ),
-) -> None:
-    """Convert files using Docling."""
-    fmts = format or _parse_env_formats() or [OutputFormat.MARKDOWN]
-    if not SETTINGS["verbose"]:
-        warnings.filterwarnings("ignore")
-    if source.startswith(("http://", "https://")):
-        results = convert_path(source, fmts)
-    else:
-        results = convert_path(Path(source), fmts)
-    if not results:
-        console.print("No new files to process.")
-
-
-@app.command()
-def validate(
-    raw: Path = typer.Argument(..., help="Path to raw document"),
-    rendered: Path | None = typer.Argument(
-        None, help="Path to converted file"
-    ),
-    fmt: Optional[OutputFormat] = typer.Option(None, "--format", "-f"),
-    prompt: Optional[Path] = typer.Option(
-        None,
-        "--prompt",
-        help="Prompt file (overrides auto-detected *.validate.prompt.yaml)",
-        callback=_validate_prompt,
-    ),
-    model: Optional[str] = typer.Option(
-        None,
-        "--model",
-        help="Model name override",
-        callback=_validate_model,
-    ),
-    base_model_url: Optional[str] = typer.Option(
-        None, "--base-model-url", help="Model base URL override"
-    ),
-    verbose: bool = typer.Option(False, "--verbose", "-v", help="Enable debug logging"),
-    log_file: Optional[Path] = typer.Option(
-        None, "--log-file", help="Write request/response details to this file"
-    ),
-) -> None:
-    """Validate converted output against the original file."""
-    logger: logging.Logger | None = None
-    log_path = log_file
-    console_local = Console()
-    if verbose or log_path is not None:
-        logger = logging.getLogger("doc_ai.validate")
-        logger.setLevel(logging.DEBUG)
-        if verbose:
-            from rich.logging import RichHandler
-
-            sh = RichHandler(console=console_local, show_time=False)
-            sh.setLevel(logging.DEBUG)
-            logger.addHandler(sh)
-        if log_path is None:
-            log_path = raw.with_suffix(".validate.log")
-        log_path.parent.mkdir(parents=True, exist_ok=True)
-        fh = logging.FileHandler(log_path)
-        fh.setLevel(logging.DEBUG)
-        logger.addHandler(fh)
-
-    if rendered is None:
-        used_fmt = fmt or OutputFormat.MARKDOWN
-        rendered = raw.with_name(raw.name + _suffix(used_fmt))
-    else:
-        used_fmt = fmt or _infer_format(rendered)
-
-    validate_doc(
-        raw,
-        rendered,
-        used_fmt,
-        prompt,
-        model,
-        base_model_url,
-        show_progress=True,
-        logger=logger,
-        console=console_local,
-    )
-
-
-@app.command()
-def analyze(
-    source: Path = typer.Argument(..., help="Raw or converted document"),
-    fmt: Optional[OutputFormat] = typer.Option(
-        None, "--format", "-f", help="Format of converted file"
-    ),
-    prompt: Path | None = typer.Option(
-        None,
-        "--prompt",
-        "-p",
-        help="Prompt file (overrides auto-detected *.analysis.prompt.yaml)",
-        callback=_validate_prompt,
-    ),
-    output: Optional[Path] = typer.Option(
-        None,
-        "--output",
-        help="Optional output file; defaults to <doc>.analysis.json",
-    ),
-    model: Optional[str] = typer.Option(
-        None,
-        "--model",
-        help="Model name override",
-        callback=_validate_model,
-    ),
-    base_model_url: Optional[str] = typer.Option(
-        None, "--base-model-url", help="Model base URL override"
-    ),
-    require_json: bool = typer.Option(
-        False,
-        "--require-structured",
-        help="Fail if analysis output is not valid JSON",
-        is_flag=True,
-    ),
-    fail_fast: bool = typer.Option(
-        True,
-        "--fail-fast/--keep-going",
-        help="Stop processing on first validation or analysis failure",
-    ),
-) -> None:
-    """Run an analysis prompt against a converted document."""
-    markdown_doc = source
-    if ".converted" not in "".join(markdown_doc.suffixes):
-        used_fmt = fmt or OutputFormat.MARKDOWN
-        markdown_doc = source.with_name(source.name + _suffix(used_fmt))
-    analyze_doc(markdown_doc, prompt, output, model, base_model_url, require_json)
-
-
-@app.command()
-def embed(
-    source: Path = typer.Argument(..., help="Directory containing Markdown files"),
-) -> None:
-    """Generate embeddings for Markdown files."""
-    build_vector_store(source)
-
-
-@app.command("pipeline")
-def pipeline(
-    source: Path = typer.Argument(..., help="Directory with raw documents"),
-    prompt: Path = typer.Option(
-        Path(".github/prompts/doc-analysis.analysis.prompt.yaml"),
-        help="Analysis prompt file",
-        callback=_validate_prompt,
-    ),
-    format: list[OutputFormat] = typer.Option(
-        None,
-        "--format",
-        "-f",
-        help="Desired output format(s) for conversion",
-    ),
-    model: Optional[str] = typer.Option(
-        None,
-        "--model",
-        help="Model name override",
-        callback=_validate_model,
-    ),
-    base_model_url: Optional[str] = typer.Option(
-        None, "--base-model-url", help="Model base URL override"
-    ),
-    fail_fast: bool = typer.Option(
-        True,
-        "--fail-fast/--keep-going",
-        help="Stop processing on first validation or analysis failure",
-    ),
-) -> None:
-    """Run the full pipeline: convert, validate, analyze, and embed."""
-    fmts = format or _parse_env_formats() or [OutputFormat.MARKDOWN]
-    convert_path(source, fmts)
-    validation_prompt = Path(
-        ".github/prompts/validate-output.validate.prompt.yaml"
-    )
-    failures: list[tuple[str, Path, Exception]] = []
-    for raw_file in source.rglob("*"):
-        if (
-            not raw_file.is_file()
-            or raw_file.suffix.lower() not in RAW_SUFFIXES
-            or any(".converted" in part for part in raw_file.parts)
-        ):
-            continue
-        md_file = raw_file.with_name(raw_file.name + _suffix(OutputFormat.MARKDOWN))
-        if md_file.exists():
-            try:
-                validate_doc(
-                    raw_file,
-                    md_file,
-                    OutputFormat.MARKDOWN,
-                    validation_prompt,
-                    model,
-                    base_model_url,
-                )
-            except Exception as exc:  # pragma: no cover - error handling
-                failures.append(("validation", raw_file, exc))
-                console.print(
-                    f"[red]Validation failed for {raw_file}: {exc}[/red]"
-                )
-                if fail_fast:
-                    break
-            try:
-                analyze_doc(
-                    md_file, prompt=prompt, model=model, base_url=base_model_url
-                )
-            except Exception as exc:  # pragma: no cover - error handling
-                failures.append(("analysis", md_file, exc))
-                console.print(
-                    f"[red]Analysis failed for {md_file}: {exc}[/red]"
-                )
-                if fail_fast:
-                    break
-    build_vector_store(source)
-    if failures:
-        console.print("[bold red]Failures encountered during pipeline:[/bold red]")
-        for step, path, exc in failures:
-            console.print(f"- {step} {path}: {exc}")
-        raise typer.Exit(code=1)
-
+# Re-export pipeline callback for tests and external use.
+from .pipeline import pipeline
 
 __all__ = [
     "app",
@@ -439,6 +193,7 @@ __all__ = [
     "interactive_shell",
     "get_completions",
     "main",
+    "pipeline",
 ]
