+++ scripts/build_vector_store.py
@@ -1,19 +1,14 @@
-import argparse
-from pathlib import Path
-
+import sys
 from dotenv import load_dotenv
+
+from doc_ai.cli import app, ENV_FILE
+
+
+def main(argv: list[str] | None = None) -> None:
+    load_dotenv(ENV_FILE)
+    args = argv if argv is not None else sys.argv[1:]
+    app(prog_name="build_vector_store.py", args=["embed", *args], standalone_mode=False)
 
 
 if __name__ == "__main__":
-    load_dotenv()
-    from doc_ai.github.vector import build_vector_store
-
-    parser = argparse.ArgumentParser()
-    parser.add_argument("source", type=Path, help="Directory containing Markdown files")
-    parser.add_argument(
-        "--fail-fast",
-        action="store_true",
-        help="Abort immediately on the first HTTP error",
-    )
-    args = parser.parse_args()
-    build_vector_store(args.source, fail_fast=args.fail_fast)
+    main()
