+++ tests/test_cli_convert_no_files.py
@@ -7,6 +7,6 @@ def test_convert_cli_reports_when_no_files(monkeypatch, tmp_path):
         return {}
     monkeypatch.setattr("doc_ai.cli.convert_path", fake_convert_path)
     runner = CliRunner()
-    result = runner.invoke(app, ["convert", str(tmp_path), "-f", "markdown"])
+    result = runner.invoke(app, ["convert", "-f", "markdown", str(tmp_path)])
     assert "No new files to process." in result.stdout
